requirements.txt
# Flask Backend Dependencies
# Install with: pip install -r requirements.txt

# Core Flask
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Authentication & Security
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Email (optional - for verification emails)
Flask-Mail==0.9.1

# Environment variables
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
